dev_requirements.txt
flask-testing==0.8.0
yapf==0.30.0
nose2==0.9.2
pylint==2.5.3
pylint-flask==0.6